Add files via upload
This is the executable for janes_window. This version runs on Linux. If you want the raw source code, that's also posted here as well, in the janes_window.py file. 
janes_window is a play on John the Ripper...John Doe...Jane Doe...get it? Anywho, the Window part was added because this hash cracking program is specific for Windows NTLM hashes, the kind you get running "hashdump" from the meterpreter console. 
The controls are pretty intuitive. You can choose to crack a file full of hashes, or enter hashes separated by commas. You can then choose from the available password list, or enter your own location. The program is specific to NTLM hashes only. 
It's a type of "wrapper" program with John the Ripper at it's core. We made this as an easy way to crack all of the windows hashes we get during our pentesting practices, CTF's, KoTH's and so on. Enjoy.